refactor(settings): extract one-shot migrations into extension
Pulls the three migrate* helpers and their default-constant enums
(WindowPaddingDefaults, BuiltInSettingsWindowDefaults) out of
BellithSettings.swift into BellithSettings+Migration.swift.

Migration code is naturally feature-isolated and grows over time —
giving it its own file keeps the main settings class focused on the
typed-property accessors.

BellithSettings.swift: 924 → 862 lines.
